Ignore namespace on cluster-scoped watched objects
Previously, it used the Namespace field in the watchedObject to build
the resource interface, but this caused problems if a namespace was
provided on a cluster-scoped object. The user might expect that the
namespace will just be ignored on these kinds of objects, like it is in
`kubectl`, so the previous behavior may be surprising.

Now, it uses the scoping information that it already had from the API
resource list.

Refs:
 - https://issues.redhat.com/browse/ACM-5547

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>